Refactor for better performance
1. Use a fixed thread to handle new lines instead of timers. A new
thread comes with a new timer, so this change avoid spawning a lot of
threads.
2. Avoid executing unfold/fold for each line when a filter is applied.
3. "adb_delay_scrolling" is removed. It is a meaningless feature since
we can toggle manual scrolling and automatic scrolling. Further more,
the implementation would become much more complicated if we add support
for it.

With this patch, loading 20000+ lines of log when connecting to a
device is much faster than before.